メール送受信のライブラリ組み込み完了、現状はピッカー選択を5個に設定

+++ MarriageCreate/SMTPLibrary/NSStream+SKPSMTPExtensions.m
@@ -47,7 +47,7 @@
     host = CFHostCreateWithName(NULL, (CFStringRef) hostName);
     if (host != NULL) 
     {
-        (void) CFStreamCreatePairWithSocketToCFHost(NULL, host, port, &readStream, &writeStream);
+        (void) CFStreamCreatePairWithSocketToCFHost(NULL, host, (int)port, &readStream, &writeStream);
         CFRelease(host);
     }
